+++ x-silo/contracts/modules/XRedeemPolicy.sol
@@ -4,6 +4,7 @@ pragma solidity 0.8.28;
 import {Ownable2Step, Ownable} from "openzeppelin5/access/Ownable2Step.sol";
 import {ERC20, IERC20} from "openzeppelin5/token/ERC20/ERC20.sol";
 import {Math} from "openzeppelin5/utils/math/Math.sol";
+import {SafeERC20} from "openzeppelin5/token/ERC20/utils/SafeERC20.sol";
 
 import {TransientReentrancy} from "silo-core/contracts/hooks/_common/TransientReentrancy.sol";
 import {IXRedeemPolicy} from "../interfaces/IXRedeemPolicy.sol";
@@ -11,6 +12,8 @@ import {IXRedeemPolicy} from "../interfaces/IXRedeemPolicy.sol";
 /// @dev based on Camelot's xGRAIL
 /// @notice Policy for redeem xSilo back to Silo
 abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentrancy {
+    using SafeERC20 for IERC20;
+
     uint256 internal constant _PRECISION = 1e18;
 
     /// @inheritdoc IXRedeemPolicy
@@ -18,6 +21,9 @@ abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentr
 
     /// @inheritdoc IXRedeemPolicy
     uint256 public constant MAX_REDEEM_DURATION_CAP = 365 days;
+
+    /// @dev total Silo amount represent sum of all user redeems items, including penalty and fees
+    uint256 public pendingLockedSilo;
 
     // Redeeming min/max settings are updatable at any time by owner.
     // Except for the max redeem ratio, which is capped at 1:1.
@@ -62,7 +68,7 @@ abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentr
         nonReentrant
         returns (uint256 siloAmountAfterVesting)
     {
-        require(_xSiloAmountToBurn > 0, ZeroAmount());
+        require(_xSiloAmountToBurn != 0, ZeroAmount());
         require(_duration >= minRedeemDuration, DurationTooLow());
         require(_duration <= maxRedeemDuration, DurationTooHi());
 
@@ -86,7 +92,12 @@ abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentr
                 })
             );
 
-            _transferShares(msg.sender, address(this), _xSiloAmountToBurn);
+            // burn shares immediately so they leave the active pool
+            // at the same time it busted other users share value
+            _burnShares(msg.sender,_xSiloAmountToBurn);
+
+            // track the owed SILO separately
+            pendingLockedSilo += siloAmountAfterVesting;
         } else {
             // immediately redeem for SILO
             _withdraw({
@@ -104,37 +115,33 @@ abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentr
         RedeemInfo storage redeem_ = _userRedeems[msg.sender][redeemIndex];
         require(block.timestamp >= redeem_.endTime, VestingNotOver());
 
-        emit FinalizeRedeem(msg.sender, redeem_.siloAmountAfterVesting, redeem_.xSiloAmountToBurn);
+        emit FinalizeRedeem(msg.sender, redeem_.siloAmountAfterVesting);
 
-        uint256 assetsToTransfer = redeem_.siloAmountAfterVesting;
-        uint256 sharesToBurn = redeem_.xSiloAmountToBurn;
+        // release the SILO owed
+        pendingLockedSilo -= redeem_.siloAmountAfterVesting;
+        // transfer the vested funds to the user
+        IERC20(_getSiloToken()).safeTransfer(msg.sender, redeem_.siloAmountAfterVesting);
 
         // remove redeem entry
         _deleteRedeemEntry(redeemIndex);
-
-        _withdraw({
-            _caller: address(this),
-            _receiver: msg.sender,
-            _owner: address(this),
-            _assetsToTransfer: assetsToTransfer,
-            _sharesToBurn: sharesToBurn
-        });
     }
 
     /// @inheritdoc IXRedeemPolicy
     function cancelRedeem(uint256 _redeemIndex) external nonReentrant validateRedeem(msg.sender, _redeemIndex) {
         RedeemInfo storage redeemCache = _userRedeems[msg.sender][_redeemIndex];
 
-        uint256 toTransfer = _convertToShares(redeemCache.currentSiloAmount, Math.Rounding.Floor);
-        uint256 toBurn = redeemCache.xSiloAmountToBurn - toTransfer;
+        uint256 shares = _convertToShares(redeemCache.currentSiloAmount, Math.Rounding.Floor);
+        require(shares != 0, CancelGeneratesZeroShares());
 
-        emit CancelRedeem(msg.sender, toTransfer, toBurn);
+        // mint back exact deposit equivalent before unmarking (i.e. deposit not yet in vault)
+        _mintShares(address(msg.sender), shares);
+        // free up SILO owed to users
+        pendingLockedSilo -= redeemCache.siloAmountAfterVesting;
+
+        emit CancelRedeem(msg.sender, redeemCache.currentSiloAmount, shares);
 
         // remove redeem entry
         _deleteRedeemEntry(_redeemIndex);
-
-        if (toTransfer != 0) _transferShares(address(this), msg.sender, toTransfer);
-        if (toBurn != 0) _burnShares(address(this), toBurn);
     }
 
     /// @inheritdoc IXRedeemPolicy
@@ -223,7 +230,6 @@ abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentr
         uint256 ratio = _calculateRatio(_duration);
         if (ratio == 0) return type(uint256).max;
 
-
         xSiloAmountIn = Math.mulDiv(_xSiloAfterVesting, _PRECISION, ratio, Math.Rounding.Ceil);
     }
 
@@ -244,7 +250,11 @@ abstract contract XRedeemPolicy is IXRedeemPolicy, Ownable2Step, TransientReentr
         uint256 _sharesToBurn
     ) internal virtual;
 
+    function _getSiloToken() internal view virtual returns (address tokenAddress);
+
     function _burnShares(address _account, uint256 _shares) internal virtual;
+
+    function _mintShares(address _account, uint256 _shares) internal virtual;
 
     function _transferShares(address _from, address _to, uint256 _shares) internal virtual;