+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - x-silo: use rounding direction in favor of protocol for cancel redeem 
 - x-silo: ensure we set `distrubutionEnd` when reset emission
 - x-silo: add restriction when redeem duration is too hi
+- x-silo: ensure cancel redeem does not give more tokens that user should receive
 
 ### [3.6.0] - 2025-05-19
 #Added